# backend/app/Http/Controllers/BookingController.php
<?php

namespace App\Http\Controllers;

use App\Enums\BookingStatus;
use App\Events\BookingCreated;
use App\Events\BookingDeleted;
use App\Events\BookingUpdated;
use App\Exceptions\BookingCancellationException;
use App\Exceptions\PendingBookingLimitExceededException;
use App\Exceptions\RefundFailedException;
use App\Http\Requests\StoreBookingRequest;
use App\Http\Requests\UpdateBookingRequest;
use App\Http\Resources\BookingResource;
use App\Models\Booking;
use App\Services\BookingService;
use App\Services\CancellationService;
use App\Services\CreateBookingService;
use App\Services\RoomService;
use App\Traits\ApiResponse;
use Illuminate\Database\QueryException;
use Illuminate\Http\JsonResponse;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Log;
use RuntimeException;

class BookingController extends Controller
{
    use ApiResponse;

    public function __construct(
        private CreateBookingService $createBookingService,
        private BookingService $bookingService,
        private CancellationService $cancellationService,
        private RoomService $roomService
    ) {}

    /**
     * Display a listing of the resource.
     */
    public function index(Request $request): JsonResponse
    {
        // No policy guard needed: getUserBookings() filters by auth()->id(),
        // ensuring users can only see their own bookings. The viewAny policy
        // is semantically "view ALL bookings" (moderator+), not "view own".
        // Use $request->user() — check_token_valid sets the request user resolver
        // directly; auth()->id() resolves via the default guard and returns null here.
        $bookings = $this->bookingService->getUserBookings($request->user()->id);

        return response()->json([
            'success' => true,
            'data' => BookingResource::collection($bookings),
        ]);
    }

    /**
     * Store a newly created resource in storage.
     *
     * Uses CreateBookingService to prevent double-booking
     * Service handles the transaction, pessimistic locking, and retry logic
     *
     * INPUT SANITIZATION:
     * - FormRequest validation rejects invalid input
     * - Booking model trait auto-purifies guest_name (HTML Purifier, not regex)
     * - Regex blacklist = 99% bypass. HTML Purifier = 0% bypass.
     */
    public function store(StoreBookingRequest $request): JsonResponse
    {
        $validated = $request->validated();
        $additionalData = [];

        foreach (['number_of_guests', 'special_requests'] as $field) {
            if (array_key_exists($field, $validated)) {
                $additionalData[$field] = $validated[$field];
            }
        }

        try {
            // Delegate to service to create booking (includes pessimistic locking)
            $booking = $this->createBookingService->create(
                roomId: $validated['room_id'],
                checkIn: $validated['check_in'],
                checkOut: $validated['check_out'],
                guestName: $validated['guest_name'],
                guestEmail: $validated['guest_email'],
                userId: $request->user()->id,
                additionalData: $additionalData
            );
        } catch (QueryException $e) {
            // Map PostgreSQL exclusion violation (SQLSTATE 23P01) to 409 Conflict.
            //
            // Why this branch must come BEFORE the RuntimeException catch:
            // QueryException → PDOException → RuntimeException (PHP class hierarchy).
            // Without an explicit QueryException catch first, a 23P01 race would
            // fall into the generic RuntimeException → 422 branch with a raw
            // PG SQLSTATE error string leaked to the client. Worse, callers
            // would see 422 (input validation) for what is actually a concurrent
            // conflict that is safe to retry.
            //
            // Mirrors the precedent in AdminBookingController::restore (line 140-145):
            // 23P01 from the exclusion constraint == 409 with a localized message.
            if ($e->getCode() === '23P01') {
                return response()->json([
                    'success' => false,
                    'message' => __('booking.create_concurrent_conflict'),
                ], 409);
            }

            // Other DB errors: opaque 500 with logging (do not leak SQLSTATE)
            Log::error('Booking creation failed (DB error): '.$e->getMessage(), [
                'user_id' => $request->user()?->id,
                'room_id' => $validated['room_id'] ?? null,
                'sqlstate' => $e->errorInfo[0] ?? null,
            ]);

            return response()->json([
                'success' => false,
                'message' => __('booking.create_error'),
            ], 500);
        } catch (PendingBookingLimitExceededException $e) {
            return response()->json([
                'success' => false,
                'message' => $e->getMessage(),
                'errors' => [
                    'code' => $e->getErrorCode(),
                ],
            ], 422);
        } catch (RuntimeException $e) {
            // Handle business errors from the service (room unavailable, not found, etc.)
            return response()->json([
                'success' => false,
                'message' => $e->getMessage(),
            ], 422);
        } catch (\Throwable $e) {
            // Log unexpected errors
            Log::error('Booking creation failed: '.$e->getMessage(), [
                'user_id' => $request->user()?->id,
                'room_id' => $validated['room_id'] ?? null,
                'exception' => class_basename($e),
            ]);

            return response()->json([
                'success' => false,
                'message' => __('booking.create_error'),
            ], 500);
        }

        // Booking is committed. Fire event for cache invalidation and notifications.
        // Event listeners run outside the transaction; a listener failure (e.g. mail
        // transport error with QUEUE_CONNECTION=sync) must NOT roll back the booking
        // or turn a successful 201 into a 500. Log and continue.
        try {
            event(new BookingCreated($booking));
        } catch (\Throwable $e) {
            Log::error('BookingCreated event listener failed after successful booking creation', [
                'booking_id' => $booking->id,
                'exception' => class_basename($e),
                'message' => $e->getMessage(),
            ]);
        }

        return response()->json([
            'success' => true,
            'message' => __('booking.created'),
            'data' => new BookingResource($booking->load('room')),
        ], 201);
    }

    /**
     * Display the specified resource.
     */
    public function show(Booking $booking): JsonResponse
    {
        // Authorize via policy
        $this->authorize('view', $booking);

        $cachedBooking = $this->bookingService->getBookingById($booking->id);

        return response()->json([
            'success' => true,
            'data' => new BookingResource($cachedBooking),
        ]);
    }

    /**
     * Update the specified resource in storage.
     *
     * Uses CreateBookingService to update, ensuring no overlap with existing bookings
     */
    public function update(UpdateBookingRequest $request, Booking $booking): JsonResponse
    {
        // Authorize update
        $this->authorize('update', $booking);

        $validated = $request->validated();
        $updateData = [
            'guest_name' => $validated['guest_name'] ?? $booking->guest_name,
            'guest_email' => $validated['guest_email'] ?? $booking->guest_email,
        ];

        if (array_key_exists('special_requests', $validated)) {
            $updateData['special_requests'] = $validated['special_requests'];
        }

        try {
            // Store original booking data BEFORE updating (to compare changes later)
            $originalBookingData = (object) $booking->toArray();

            // Call service to update — convert date strings to Carbon instances
            // Pass guest_name and guest_email via additionalData
            $booking = $this->createBookingService->update(
                booking: $booking,
                checkIn: \Carbon\Carbon::createFromFormat('Y-m-d', $validated['check_in'])->startOfDay(),
                checkOut: \Carbon\Carbon::createFromFormat('Y-m-d', $validated['check_out'])->startOfDay(),
                additionalData: $updateData
            );

            // Dispatch event for notification - pass actual Booking model and original data
            event(new BookingUpdated($booking, $originalBookingData));

            return response()->json([
                'success' => true,
                'message' => __('booking.updated'),
                'data' => new BookingResource($booking->load('room')),
            ]);
        } catch (RuntimeException $e) {
            return response()->json([
                'success' => false,
                'message' => $e->getMessage(),
            ], 422);
        } catch (\Throwable $e) {
            Log::error('Booking update failed: '.$e->getMessage(), [
                'booking_id' => $booking->id,
                'exception' => class_basename($e),
            ]);

            return response()->json([
                'success' => false,
                'message' => __('booking.update_error'),
            ], 500);
        }
    }

    /**
     * Remove the specified resource from storage (soft delete).
     *
     * Uses soft delete to preserve audit trail - booking data is NOT permanently removed.
     * Records who deleted the booking and when for compliance.
     *
     * Regular users: Can only soft delete their own bookings
     * Admins: Can soft delete any booking
     */
    public function destroy(Request $request, Booking $booking): JsonResponse
    {
        // Authorize deletion
        $this->authorize('delete', $booking);

        // Dispatch event for cache invalidation BEFORE deleting
        // This avoids Laravel trying to restore a deleted model from the event
        event(new BookingDeleted($booking));

        // Soft delete booking with audit trail (records deleted_by)
        $this->bookingService->softDelete($booking, $request->user()->id);

        return response()->json([
            'success' => true,
            'message' => __('booking.deleted'),
        ], 200);
    }

    /**
     * Confirm a pending booking.
     *
     * Changes booking status from 'pending' to 'confirmed' and triggers
     * a queued confirmation email notification to the guest.
     *
     * Authorization: Only admins can confirm bookings
     * Rate limiting: Max 5 confirmation emails per user per minute
     */
    public function confirm(Booking $booking): JsonResponse
    {
        $this->authorize('confirm', $booking);

        if ($booking->status !== BookingStatus::PENDING) {
            return response()->json([
                'success' => false,
                'message' => __('booking.confirm_invalid_status', ['status' => $booking->status->value]),
            ], 422);
        }

        try {
            $booking = $this->bookingService->confirmBooking($booking);

            return response()->json([
                'success' => true,
                'message' => __('booking.confirmed'),
                'data' => new BookingResource($booking->load('room')),
            ]);
        } catch (\RuntimeException $e) {
            return response()->json([
                'success' => false,
                'message' => $e->getMessage(),
            ], 422);
        }
    }

    /**
     * Cancel a booking with optional refund.
     *
     * This endpoint handles the complete cancellation flow:
     * 1. Validates cancellation eligibility (ownership, status, timing)
     * 2. Calculates refund amount based on cancellation policy
     * 3. Processes refund via Stripe (if payment exists)
     * 4. Updates booking status
     * 5. Sends cancellation notification
     *
     * Authorization: Users can cancel their own bookings, admins can cancel any
     * Idempotent: Re-cancelling an already cancelled booking returns success
     */
    public function cancel(Request $request, Booking $booking): JsonResponse
    {
        $this->authorize('cancel', $booking);

        try {
            $booking = $this->cancellationService->cancel(
                $booking,
                $request->user()
            );

            // Build response message based on refund status
            $message = $this->buildCancellationMessage($booking);

            return response()->json([
                'success' => true,
                'message' => $message,
                'data' => new BookingResource($booking->load('room')),
            ]);

        } catch (BookingCancellationException $e) {
            return response()->json([
                'success' => false,
                'message' => $e->getMessage(),
                'error' => $e->getErrorCode(),
            ], $e->getHttpStatusCode());

        } catch (RefundFailedException $e) {
            return response()->json([
                'success' => false,
                'message' => __('booking.refund_failed'),
                'error' => 'refund_failed',
                'retryable' => $e->isRetryable(),
            ], $e->getHttpStatusCode());
        }
    }

    /**
     * Build appropriate cancellation message based on refund status.
     */
    private function buildCancellationMessage(Booking $booking): string
    {
        if ($booking->refund_amount && $booking->refund_amount > 0) {
            $formattedAmount = number_format($booking->refund_amount / 100, 2);

            return (string) __('booking.cancel_with_refund', ['amount' => '$'.$formattedAmount]);
        }

        if ($booking->payment_intent_id && $booking->refund_amount === 0) {
            return (string) __('booking.cancel_no_refund');
        }

        return (string) __('booking.cancelled');
    }
}